change how deps are managed
The way dependencies were managed isn't the way which is recommended
by pipenv. Dependencies shouldn't be pinned to any version unless we
want to really pin it to a specific version for specific reasons.
Correct versions to install are in the lockfile. @dependabot helps
us to keep the lockfile updated, preparing PRs for each library when
it gets a new version.